app/requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
python-multipart==0.0.20

# Agentic framework
langgraph==0.2.65
langchain-core==0.3.30
langchain-google-genai==2.0.9
langchain-community==0.3.14

# PDF parsing
PyMuPDF==1.25.3
pdfplumber==0.11.4

# Embeddings & Vector DB
chromadb==0.6.3
rank-bm25==0.2.2

# Optional: uncomment for local embeddings (adds ~2GB PyTorch)
# sentence-transformers==3.3.1

# LLM providers
google-generativeai==0.8.4

# Optional: uncomment for RAGAS evaluation
# ragas==0.2.14
# datasets==3.2.0

# Utilities
httpx==0.28.1
tenacity==9.0.0
Pillow==11.1.0
numpy>=1.26.0,<2.0.0
